requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.31.0
chromadb==0.4.22
langchain==0.1.4
langchain-community==0.0.16
sentence-transformers==2.3.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
